MADE: Move detections list into a separate file
